+++ CHANGES.md
@@ -28,7 +28,8 @@ Beta Releases
                 }
             });
     * `Label.computeScreenSpacePosition` now requires the current scene state as a parameter.
-            
+    * Passing `undefined` to any of the set functions on `Label` now throws an exception.
+
 * All `Quaternion` operations now have static versions that work with any objects exposing `x`, `y`, `z` and `w` properties.
 * Added support for nested polygons with holes. See `Polygon.configureFromPolygonHierarchy`.
 * Added `writeTextToCanvas` function which handles sizing the resulting canvas to fit the desired text.